# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rest API for catalog",
        "version": "1"
    },
    "paths": {
...
        "/api/auth/companies/building/{id}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get list of companies by building",
                "description": "Returns list of companies by building",
                "operationId": "getByBuilding",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "building id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/auth/companies/rubric/{id}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get list of companies by rubric",
                "description": "Returns list of companies by rubric",
                "operationId": "getCompanyByRubric",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "rubric id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/auth/companies/map/point/{latitude}/{longitude}/{radius}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get list of companies by map point and radius",
                "description": "Returns list of companies by map point and radius",
                "operationId": "getCompanyByMapPoint",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "path",
                        "description": "latitude",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "47.897709"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "path",
                        "description": "longitude",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "40.082366"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "path",
                        "description": "radius (km)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/auth/companies/rectangle/{latitude1}/{longitude1}/{latitude2}/{longitude2}": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get list of companies by rectangle",
                "description": "Returns list of companies by rectangle",
                "operationId": "getCompanyByRectangle",
                "parameters": [
                    {
                        "name": "latitude1",
                        "in": "path",
                        "description": "latitude1",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "47.897709"
                        }
                    },
                    {
                        "name": "longitude1",
                        "in": "path",
                        "description": "longitude1",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "40.082366"
                        }
                    },
                    {
                        "name": "latitude2",
                        "in": "path",
                        "description": "latitude2",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "48.897709"
                        }
                    },
                    {
                        "name": "longitude2",
                        "in": "path",
                        "description": "longitude2",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "41.082366"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}